chore: Solving comments
-move `handleWebhook` to `whatsapp.service`
